refactor: components/preview/ → components/previews/ 폴더명 변경 및 import 경로 반영
