+++ src/feature/CardRegister/components/previews/CardExpiryDateDisplay/CardExpiryDateDisplay.tsx
