+++ src/feature/CardRegister/components/previews/CardNumberDisplay/CardNumberDisplay.tsx
